requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Computer Vision
opencv-python-headless>=4.8.0
albumentations>=1.3.0
pillow>=10.0.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Google Cloud (for Vertex AI)
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.40.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional (for serving)
fastapi>=0.100.0
uvicorn>=0.23.0

# Visualization (optional)
matplotlib>=3.7.0
